New Version
Fixed some errors, added code to specify "internal" or "external" angle and to plot dispersion vs angle or in-plane wavevector. 

Missing a lot of the examples from the first version. There is one file 'dispersionExample.m' that will be useful though to get started. 


Internal angle: this the angle of incidence within semi-infinite medium 1. 

External angle (SPECIFIC TO RIGHT ANGLE PRISMS): I made this code for model Kretschmann-Raether plasmonics, where we excite surface plasmons by illuminating the "backside" of a metal thin film through a prism. I use either semi-hemisphere prisms, in which case the angle of illumination in air is equal to the angle inside in the prism, or I use right angle prisms. The external angle allows specification of the angle of illumination in air as it enters the prism. It's assumed that the thin films are on the hypotenuse of the prism. Light enters and exits the other two faces. All angles except 45° will refract as they enter the prism, thus the external and internal angles are different. So, for this type of modeling, one specifies the semi-infinite medium 1 to be that of the prism (i.e., ignoring that the light passes through air first, and thus not accounting for reflections off this initial interface) and then can choose to specify angles of incidence in the air (external) or in the glass (internal) and can also choose to plot dispersion vs. external or internal angles.